+++ game-frontend/projects/game-admin/src/app/components/items-filter/items-filter.component.ts
@@ -1,5 +1,5 @@
 import { Component, EventEmitter, Input, OnChanges, Output } from '@angular/core';
-import { WidgetFilter } from '@owge/widgets';
+import { WidgetFilter } from '@owge/core';
 import { AdminFactionService } from '../../services/admin-faction.service';
 
 @Component({