+++ game-frontend/projects/game-admin/src/app/services/admin-special-location.service.ts
@@ -1,9 +1,9 @@
 import { Injectable } from '@angular/core';
-import { validContext } from '@owge/core';
+import { validContext, WidgetFilter } from '@owge/core';
 import {
-    AbstractCrudService, CrudConfig, CrudServiceAuthControl, SpecialLocation, UniverseGameService, WithImprovementsCrudMixin
+    AbstractCrudService, CrudConfig, CrudServiceAuthControl, SpecialLocation,
+    UniverseGameService, WithImprovementsCrudMixin, WidgetFilterUtil
 } from '@owge/universe';
-import { WidgetFilter, WidgetFilterUtil } from '@owge/widgets';
 import { take } from 'rxjs/operators';
 import { Mixin } from 'ts-mixer';