+++ game-frontend/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 v0.10.2 (latest)
 =================
+* [class=Admin] __Fix:__ [Requirements groups not working in admin panel #419](https://github.com/KevinGuancheDarias/owge/issues/419)  
 * __Improvement:__ [Display unit type in the unit's screen #423](https://github.com/KevinGuancheDarias/owge/issues/423)  
 * __Fix:__ [Attackable types in info modal of unit doesn't display the colors, nor in classic, neither in neon #420](https://github.com/KevinGuancheDarias/owge/issues/420)  
 * __Fix:__ [Neón doesn't display the unit upgrade requirements colors #421](https://github.com/KevinGuancheDarias/owge/issues/421)